Add operator console, CI and toolchain; fix criticals found by adversarial audit
Two pieces of work: the FAVL Command Center console, and a repository-wide
hardening pass driven by a 12-agent adversarial audit that confirmed 23
findings and refuted 13.

apps/console — operator console
- Next.js 15 / React 19 / Tailwind v4, pinned to the patched 15.5.22 with
  postcss and sharp overridden to clear three high-severity advisories
- design tokens carry state and nothing else: there is no brand accent, so
  any colour on screen means something. Steady state is still; only state
  transitions animate, and prefers-reduced-motion is honoured
- command palette, sidebar collapsing to icons, contextual Inspector,
  invocation Timeline, status bar
- wired to real data only: agents, connectors, invocations, health, metrics.
  The other ten sections appear disabled with the milestone that delivers
  each. Nothing in the console is mock data
- the browser never holds a credential: route handlers attach a server-side
  token and call through APISIX, so the console exercises the real
  authenticated path rather than reaching services directly

Criticals fixed
- SSRF privilege escalation: allow_private_addresses, allowed_schemes and
  allow_plaintext_acknowledged were read from the caller-supplied connector
  config. Any principal who could create a connector could point a hostname
  at RFC1918 space, set the flag, and turn the gateway into an authenticated
  proxy into the internal network — Postgres, NATS, Keycloak and the APISIX
  control port. A destination must never vouch for itself. These are now
  operator-only environment settings; connector config may narrow the policy
  and never widen it, and the keys are rejected at the API edge
- SSRF via embedded IPv4: ::ffff:169.254.169.254 reaches the metadata service
  but is not link-local on Python 3.11, which is what the containers run.
  Tests ran on 3.13, where the stdlib does unwrap it, so the suite passed
  while production was exposed. Addresses are now normalised before
  classification (v4-mapped, 6to4, Teredo, NAT64) and forbidden ranges are
  declared as explicit networks rather than inferred from is_* properties
  that change between releases
- outbox partial index compiled to a constant-false predicate:
  func.lower("status") produced a bound literal, so the ORM-declared index
  was `WHERE lower('status') = 'pending'` — permanently empty and never
  matched to the claim query
- OutboxStalledWhileWriting could never fire: `and` joined operands with
  different label sets, and the rate() series does not exist when a publisher
  has never succeeded — precisely the case being alerted on. Rewritten with
  `unless`, which is empty-safe

Security and correctness
- tenant and actor now derive from the gateway-verified token, never from a
  client header; APISIX strips inbound X-Tenant-ID/X-Actor-ID and the service
  fails closed without verified claims. Invocation reads are tenant-scoped,
  returning 404 rather than 403 so an id in another tenant is not confirmed
- /internal/* requires a service credential compared in constant time;
  network position is not authentication
- OIDC discovery and JWKS now verify TLS on all routes
- connector config is redacted on every outbound path, so a credential that
  slips past validation is not republished through the API or into an event
- accept() captures ORM attributes before flush: reading them in the
  IntegrityError branch touched an identity map that rollback had expired,
  raising MissingGreenlet on the idempotency race path
- agent version now increments per invocation; every agent.invoked event
  previously carried version 1 and this repo's own consumer ordering rule
  discarded all but the first
- invoke_agent releases its database connection before fanning out, instead
  of holding one of fifteen across N external calls of up to 15s
- replayed failures return the original attempt's status; a replay of a 502
  previously returned 200, so a caller retrying after a failure would
  conclude it had succeeded
- public /health/ready reduced to booleans; dependency exception strings and
  outbox counters moved to the cluster-internal /readyz

Frontend
- ClientClock defers timezone-dependent formatting past hydration
- the unauthenticated POST /api/agents proxy is removed: route handlers carry
  a gateway credential, so a write endpoint without an origin check is a
  confused deputy. lib/origin.ts guards any future write
- command palette is a real dialog: role, aria-modal, working Escape, focus
  restored to the opener. Table rows are keyboard-operable
- connector status carries a text label, not colour alone

Toolchain and supply chain
- ruff (incl. bandit and async rules) and mypy --strict both clean across 36
  source files; `make check` runs lint, types and tests in CI order
- GitHub Actions: Python gate, console gate with production-only npm audit,
  compose and promtool validation, Trivy image scanning with SARIF upload,
  SBOM generation, gitleaks history scan, and the outbox suite on push
- scripts/check_alert_metrics.py fails the build when an alert references a
  metric no code path emits; verified against an injected typo
- all container images pinned by tag@digest, including the floating `latest`
  on the NATS exporter, with scripts/update_image_digests.sh to refresh
- Kubernetes workloads added: the /livez and /readyz split had nothing
  consuming it. Startup, liveness and readiness probes, migrations moved to a
  PreSync Job so replicas cannot race the Alembic lock, PDBs, topology
  spread, non-root with a read-only root filesystem, resource bounds
- tests run from a repository-local .venv with PYTHONPATH cleared

Verification
- 162 offline tests, ruff and mypy --strict clean
- tests/verify_outbox.sh 15/15 including four SIGKILLs under load
- tests/verify_identity.sh 4/4: a forged X-Userinfo leaks nothing, an
  unauthenticated read is refused, results span exactly one tenant
- full stack rebuilt from pinned digests; console verified against live data

+++ packages/favl-outbox/favl_outbox/timing.py
@@ -28,7 +28,7 @@ class RetryPolicy:
     def backoff_for(self, attempt: int) -> float:
         """Upper bound of the delay scheduled after `attempt` failures."""
         raw = min(self.base_seconds * (2 ** max(attempt - 1, 0)), self.cap_seconds)
-        return raw * (1 + self.jitter_ratio)
+        return float(raw * (1 + self.jitter_ratio))
 
 
 @dataclass(frozen=True)
@@ -56,9 +56,7 @@ class DuplicateWindowTooSmall(RuntimeError):
     """Configuration would let a retry escape deduplication."""
 
 
-def worst_case_retry_horizon(
-    policy: RetryPolicy, delays: OperationalDelays
-) -> float:
+def worst_case_retry_horizon(policy: RetryPolicy, delays: OperationalDelays) -> float:
     """Longest possible span from first publish attempt to final attempt.
 
     Every attempt pays the operational cost; every gap between attempts pays
